feat: api route for rating

+++ src/pages/api/book/delete.ts
@@ -6,10 +6,27 @@ export default async function handler(
   res: NextApiResponse
 ) {
   if (req.method !== "DELETE") {
-    res.status(501).json("Method not implemented")
+    res.status(501).json("Method not implemented.")
+    return
   }
 
   const id = req.body.id
+
+  if (!id) {
+    res.status(400).json({ message: "Id not found." })
+    return
+  }
+
+  const book = await prisma.book.findUnique({
+    where: {
+      id,
+    },
+  })
+
+  if (book === null) {
+    res.status(404).json({ message: "Book not found." })
+    return
+  }
 
   const deletedBook = await prisma.book.delete({
     where: {
